common: restructure into lib files
