+++ sonar-common/lib/symbols.js
@@ -1,0 +1,11 @@
+const RECORD = Symbol('sonar-record')
+const TYPE = Symbol('sonar-type')
+const FIELD = Symbol('sonar-field')
+const SC = Symbol('sonar-schema')
+
+module.exports = {
+  RECORD,
+  TYPE,
+  SC,
+  FIELD
+}
